+++ WebAPI/Controllers/StationsController.cs
@@ -4,6 +4,7 @@ using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
 using WebApi.Entities;
 using WebApi.Interface;
+using WebApi.Services;
 
 [ApiController]
 [Route("api/v1/[controller]/[action]")]
@@ -11,13 +12,16 @@ public class StationsController : ControllerBase
 {
     private IStationsService _stationsService;
     private IMapper _mapper;
+    private IFareChartService _fareChartService;
 
     public StationsController(
         IStationsService stationsService,
-        IMapper mapper)
+        IMapper mapper,
+        IFareChartService fareChartService)
     {
         _stationsService = stationsService;
         _mapper = mapper;
+        _fareChartService = fareChartService;
     }
 
     [HttpGet]
@@ -69,6 +73,11 @@ public class StationsController : ControllerBase
     [HttpDelete("{id}")]
     public IActionResult Delete(int id)
     {
+        if (_fareChartService.GetAll().Any(e => e.FromStationId == id || e.ToStationId == id))
+        {
+            return BadRequest(new { message = "This Chart Info already used in Fare Chart." });
+
+        }
         _stationsService.Delete(id);
         return Ok(new { message = "Stations deleted" });
     }
